requirements.txt
python-telegram-bot==13.7
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5

